app/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.2.1 
SQLAlchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0 
Flask>=2.3
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
Flask-Cors>=4.0 